requirements.txt
openai-whisper
ffmpeg-python
pathlib
yt-dlp